added owner id to reservation gorm impl delete function

+++ pkg/repositories/gormimpl/reservation_test.go
@@ -44,12 +44,12 @@ func TestDelete(t *testing.T) {
 	expectedReservation := GetReservation()
 
 	GlobalMock.NewMock().WithQuery(
-		`DELETE FROM "reservations" WHERE "reservations"."dataset_project" = $1 AND "reservations"."dataset_name" = $2 AND "reservations"."dataset_domain" = $3 AND "reservations"."dataset_version" = $4 AND "reservations"."tag_name" = $5`,
+		`DELETE FROM "reservations" WHERE "reservations"."dataset_project" = $1 AND "reservations"."dataset_name" = $2 AND "reservations"."dataset_domain" = $3 AND "reservations"."dataset_version" = $4 AND "reservations"."tag_name" = $5 AND "reservations"."owner_id" = $6`,
 	).WithRowsNum(1)
 
 	reservationRepo := getReservationRepo(t)
 
-	err := reservationRepo.Delete(context.Background(), expectedReservation.ReservationKey)
+	err := reservationRepo.Delete(context.Background(), expectedReservation.ReservationKey, expectedReservation.OwnerID)
 	assert.NoError(t, err)
 }
 
@@ -59,12 +59,12 @@ func TestDeleteFailure(t *testing.T) {
 	expectedReservation := GetReservation()
 
 	GlobalMock.NewMock().WithQuery(
-		`DELETE FROM "reservations" WHERE "reservations"."dataset_project" = $1 AND "reservations"."dataset_name" = $2 AND "reservations"."dataset_domain" = $3 AND "reservations"."dataset_version" = $4 AND "reservations"."tag_name" = $5`,
+		`DELETE FROM "reservations" WHERE "reservations"."dataset_project" = $1 AND "reservations"."dataset_name" = $2 AND "reservations"."dataset_domain" = $3 AND "reservations"."dataset_version" = $4 AND "reservations"."tag_name" = $5 AND "reservations"."owner_id" = $6`,
 	).WithRowsNum(0)
 
 	reservationRepo := getReservationRepo(t)
 
-	err := reservationRepo.Delete(context.Background(), expectedReservation.ReservationKey)
+	err := reservationRepo.Delete(context.Background(), expectedReservation.ReservationKey, expectedReservation.OwnerID)
 	assert.Error(t, err)
 	assert.Equal(t, "missing entity of type Reservation with identifier dataset_id:<project:\"testProject\" name:\"testDataset\" domain:\"testDomain\" version:\"testVersion\" > tag_name:\"testTag\" ", err.Error())
 }
